v2.4.7.4 -> v2.4.7.5
  - Alan Cox: more merging
  - L.C. Chang: new SiS IDE PCI id's.
  - Maciej Rozycki: make MP table parsing more anal. Should fix broken P4 MP tables.
  - Leonard Zubkoff: merge DAC960 completion changes
  - Christoph Rohland: saner tmpfs mount-time limit behaviour (and remount)
  - me: buffer.c logic update - faster and hopefully livelock-free

+++ drivers/net/pcmcia/ibmtr_cs.c
@@ -49,7 +49,7 @@
 #include <linux/init.h>
 #include <linux/sched.h>
 #include <linux/ptrace.h>
-#include <linux/malloc.h>
+#include <linux/slab.h>
 #include <linux/string.h>
 #include <linux/timer.h>
 #include <linux/module.h>
@@ -361,7 +361,7 @@ static void ibmtr_config(dev_link_t *link)
     mem.CardOffset = mmiobase;
     mem.Page = 0;
     CS_CHECK(MapMemPage, link->win, &mem);
-    ti->mmio = (u_long)ioremap(req.Base, req.Size);
+    ti->mmio = ioremap(req.Base, req.Size);
 
     /* Allocate the SRAM memory window */
     req.Attributes = WIN_DATA_WIDTH_16|WIN_MEMORY_TYPE_CM|WIN_ENABLE;
